tests: rename current e2e test to integrated test
This is not true e2e test
