+++ src/tests/integrated/bulletin.py
