+++ src/tests/integrated/manage/pro/specialscore.py
@@ -1,7 +1,7 @@
 import os
 import json
 
-from tests.e2e.util import AsyncTest, AccountContext
+from tests.integrated.util import AsyncTest, AccountContext
 from services.chal import ChalConst, Compiler
 from services.pro import ProConst, CheckerType, SummaryType
