+++ src/tests/integrated/manage/pro/update.py
@@ -1,7 +1,7 @@
 import os
 import json
 
-from tests.e2e.util import AsyncTest, AccountContext
+from tests.integrated.util import AsyncTest, AccountContext
 from services.pro import ProConst, CheckerType
 from services.chal import Compiler
