+++ src/tests/integrated/manage/pro/updatetests.py
@@ -4,7 +4,7 @@ import copy
 import json
 
 from services.chal import ChalConst
-from tests.e2e.util import AsyncTest, AccountContext
+from tests.integrated.util import AsyncTest, AccountContext
 
 class ManageProUpdateTestsTest(AsyncTest):
     async def _upload_file(self, filepath, session):
